+++ src/main/java/com/bae/service/ManagerService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.bae.exceptions.ManagerNotFoundException;
 import com.bae.persistance.domain.Manager;
 import com.bae.persistance.repository.ManagerRepository;
 
@@ -29,6 +30,9 @@ public class ManagerService {
 	}
 	
 	public String deleteManager(Long managerID) {
+		if (!this.managerRepo.existsById(managerID)) {
+			throw new ManagerNotFoundException();
+		}
 		managerRepo.deleteById(managerID);
 		return "Manager successfully deleted";
 	}
